Refactor logic into its own component
